Prototype: Basic Building Mechanics
Clicking on a pre-existing blocks to see the potential places to build
another tetrimino. Clicking on one of these potential places previews
how the tetrimino will look in that position. Use two-touch gestures to
rotate the tetrimino. Click the tetrimino to build it in the place it is
currently being previewed. Tetrimino preview is colored red when it is
unable to be placed there due to collisions with other pre-existing
blocks. Colored green otherwise.