Better Prioritization of Climbed Blocks
Now, every time a block is climbed, the lemming increases the priority
of all blocks they have climbed, where the blocks' priorities are
increased based on their level. The higher the block, the higher the
priority increase.

Causes a sort of trickle down effect which speeds up the path-finding
process considerably.